requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
aiosqlite==0.20.0
pandas==2.2.3
pydantic-settings==2.7.1
firecrawl-py==1.16.0
httpx==0.28.1
